+++ src/app/profile/change-password/page.tsx
@@ -1,25 +1,12 @@
 import type { Metadata } from 'next';
-import { redirect, notFound } from 'next/navigation';
 import { getProfile } from '@/app/profile/data/get-profile';
 import { ChangePassword } from '@/app/profile/change-password/change-password';
-import { getSession } from '@/lib/session';
 
 export const metadata: Metadata = {
     title: 'Change Password',
 };
 
 export default async function ChangePasswordPage() {
-    const session = await getSession();
-
-    if (!session) {
-        redirect('/account/login');
-    }
-
-    const profile = await getProfile(session.sub);
-
-    if (!profile || profile.isLockedOut) {
-        notFound();
-    }
-
+    const profile = await getProfile();
     return <ChangePassword profile={profile} />;
 }
