+++ src/app/profile/data/get-profile.ts
@@ -1,13 +1,19 @@
 import 'server-only';
 
+import { cache } from 'react';
+import { notFound, redirect } from 'next/navigation';
 import { eq, sql } from 'drizzle-orm';
-import type { GetProfileResponse } from '@/app/profile/profile-types';
 import { db } from '@/db';
 import { users } from '@/db/schema/users';
+import { getSession } from '@/lib/session';
 
-export async function getProfile(
-    userId: string
-): Promise<GetProfileResponse | undefined> {
+export const getProfile = cache(async () => {
+    const session = await getSession();
+
+    if (!session) {
+        redirect('/account/login');
+    }
+
     const [user] = await db
         .select({
             id: users.id,
@@ -19,11 +25,11 @@ export async function getProfile(
             version: sql<number>`"xmin"`.mapWith(Number),
         })
         .from(users)
-        .where(eq(users.id, userId))
+        .where(eq(users.id, session.sub))
         .limit(1);
 
-    if (!user) {
-        return undefined;
+    if (!user || user.isLockedOut) {
+        notFound();
     }
 
     return {
@@ -32,6 +38,7 @@ export async function getProfile(
         firstName: user.firstName,
         lastName: user.lastName,
         dateOfBirth: user.dateOfBirth,
+        isLockedOut: user.isLockedOut,
         version: user.version,
     };
-}
+});
