+++ src/app/profile/page.tsx
@@ -1,25 +1,12 @@
 import type { Metadata } from 'next';
-import { notFound, redirect } from 'next/navigation';
 import { getProfile } from '@/app/profile/data/get-profile';
 import { Profile } from '@/app/profile/profile';
-import { getSession } from '@/lib/session';
 
 export const metadata: Metadata = {
     title: 'My Account',
 };
 
 export default async function ProfilePage() {
-    const session = await getSession();
-
-    if (!session) {
-        redirect('/account/login');
-    }
-
-    const profile = await getProfile(session.sub);
-
-    if (!profile || profile.isLockedOut) {
-        notFound();
-    }
-
+    const profile = await getProfile();
     return <Profile profile={profile} />;
 }
